Restructuring camera controllers
