+++ Assets/Scripts/CameraController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: bcf15fdc3f0056641b482e739c9c7815
+guid: d113ce88fdb5f9b4e91573dcdd430c45
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
