+++ Assets/Scripts/OrbitViewManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f9bff4d4b5d44484caee2da89444adfd
+guid: 09626a0e75e0c6c4b92087416f197c80
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
